refactor(front): move skills DTO types out of lib/api into types/api
Sixth batch of the lib/api DTO migration: internal skills endpoints. Transport
DTO types and zod schemas move to the mirrored path under types/api/*, leaving
business logic in place.

- 4 pure-contract files moved wholesale: skills/index, skills/editors,
  assistant/skills/history, assistant/skills/suggestions.
- 2 files trimmed (DTO extracted, logic stays): skills/description_suggestion
  (drops unused zod import; keeps getSkillDescriptionSuggestion),
  skills/existing_skill_checker (keeps getSimilarSkills).
- 16 consumers updated; bare-directory import "@app/lib/api/skills" (index.ts)
  repointed to "@app/types/api/skills"; one mixed-import builder file split.

skills/detection/github/import_skills.ts is intentionally excluded: it has a
public v1 consumer and belongs in the SDK, handled in a later pass.
