+++ front-api/routes/w/[wId]/builder/skills/suggestions.ts
@@ -1,12 +1,10 @@
-import {
-  getSkillDescriptionSuggestion,
-  PostSkillSuggestionsRequestBodySchema,
-} from "@app/lib/api/skills/description_suggestion";
+import { getSkillDescriptionSuggestion } from "@app/lib/api/skills/description_suggestion";
+import { PostSkillSuggestionsRequestBodySchema } from "@app/types/api/skills/description_suggestion";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { apiError } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 
-export type { PostSkillSuggestionsRequestBody } from "@app/lib/api/skills/description_suggestion";
+export type { PostSkillSuggestionsRequestBody } from "@app/types/api/skills/description_suggestion";
 
 // Mounted at /api/w/:wId/builder/skills/suggestions.
 const app = workspaceApp();
