+++ front-api/routes/w/[wId]/skills/reinforcement_spend.ts
@@ -1,7 +1,7 @@
-import type { GetSkillsSpendResponseBody } from "@app/lib/api/skills";
 import { getCurrentPeriod } from "@app/lib/reinforcement/billing";
 import { SelfImprovingSkillsUsageResource } from "@app/lib/resources/self_improving_skills_usage_resource";
 import { SkillResource } from "@app/lib/resources/skill/skill_resource";
+import type { GetSkillsSpendResponseBody } from "@app/types/api/skills";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
