+++ front/lib/swr/skill_configurations.ts
@@ -1,19 +1,19 @@
 import type { ImportFormValues } from "@app/components/skills/import/formSchema";
 import { useDebounceWithAbort } from "@app/hooks/useDebounce";
 import { useSendNotification } from "@app/hooks/useNotification";
-import type { GetSkillHistoryResponseBody } from "@app/lib/api/assistant/skills/history";
-import type {
-  GetSkillResponseBody,
-  GetSkillsWithRelationsResponseBody,
-  GetSkillWithRelationsResponseBody,
-} from "@app/lib/api/skills";
 import type { ImportSkillsResponseBody } from "@app/lib/api/skills/detection/github/import_skills";
-import type { GetSimilarSkillsResponseBody } from "@app/lib/api/skills/existing_skill_checker";
 import type {
   DetectedSkillSummary,
   DetectSkillsResponseBody,
 } from "@app/lib/skill_detection";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
+import type { GetSkillHistoryResponseBody } from "@app/types/api/assistant/skills/history";
+import type {
+  GetSkillResponseBody,
+  GetSkillsWithRelationsResponseBody,
+  GetSkillWithRelationsResponseBody,
+} from "@app/types/api/skills";
+import type { GetSimilarSkillsResponseBody } from "@app/types/api/skills/existing_skill_checker";
 import type {
   SkillReinforcementMode,
   SkillStatus,
