+++ front/types/api/assistant/skills/history.ts
