+++ front/types/api/assistant/skills/suggestions.ts
