+++ front/types/api/skills/description_suggestion.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+export const PostSkillSuggestionsRequestBodySchema = z.object({
+  instructions: z.string(),
+  agentFacingDescription: z.string(),
+  tools: z.array(z.object({ name: z.string(), description: z.string() })),
+});
+export type PostSkillSuggestionsRequestBody = z.infer<
+  typeof PostSkillSuggestionsRequestBodySchema
+>;
