+++ front/types/api/skills/editors.ts
