+++ front/types/api/skills/existing_skill_checker.ts
@@ -1,0 +1,3 @@
+export type GetSimilarSkillsResponseBody = {
+  similar_skills: string[];
+};
