+++ front/types/api/skills/index.ts
